ipod: Fix bug for iPod 3G fitted with CF adapters
The CF adapter is identified as a disk with model FC-1307 SD to CF
Adapter.

HAVE_ATA_POWER_OFF is undefined because it is hypothesised that the
adapter does not handle request to cycle power gracefully. IDENTIFY
reports Power Management as 'Unsupported' for this CF adapter.

Even with HAVE_ATA_POWER_OFF undefined, there was still a kernel crash.
By undefining HAVE_DISK_STORAGE, the crash was resolved.

It would be nice to identify the real causes of the crash, but I am
quite happy with these changes for now.